+++ app/helpers/admin/resources/data_types/date_helper.rb
@@ -1,9 +1,13 @@
 module Admin::Resources::DataTypes::DateHelper
 
   def date_filter(filter)
-    values = %w(today last_few_days last_7_days last_30_days)
+    values = { :all_day => "Today",
+               :all_week => "This week",
+               :all_month => "This month",
+               :all_year => "This year" }
+
     items = [[@resource.human_attribute_name(filter).capitalize, ""]]
-    items += values.map { |v| [Typus::I18n.t(v.humanize), v] }
+    items += values.map { |k, v| [Typus::I18n.t(v), k] }
   end
 
   alias_method :datetime_filter, :date_filter